revert: the latest changes were causing memory leaks and slowing dev
