--- packages/ecc-utils-design/scripts/make-metadata.js
@@ -1,16 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-
-//
-// This script runs the Custom Elements Manifest analyzer to generate custom-elements.json
-//
-
-import { execSync } from "child_process";
-import { program } from "commander";
-
-program.option("-o, --outdir <string>").parse();
-const options = program.opts();
-
-execSync(
-  `custom-elements-manifest analyze --litelement --outdir "${options.outdir}" `,
-  { stdio: "inherit" }
-);
