+++ packages/ecc-utils-design/src/components/collection/index.ts
@@ -1,15 +1,12 @@
-import EccUtilsDesignCollection from "./collection.js";
+import Collection from "./collection.js";
 
 export * from "./collection.js";
-export default EccUtilsDesignCollection;
+export default Collection;
 
-window.customElements.define(
-  "ecc-utils-design-collection",
-  EccUtilsDesignCollection
-);
+window.customElements.define("ecc-utils-design-collection", Collection);
 
 declare global {
   interface HTMLElementTagNameMap {
-    "ecc-utils-design-collection": EccUtilsDesignCollection;
+    "ecc-utils-design-collection": Collection;
   }
 }
