+++ packages/ecc-utils-design/src/components/form/index.ts
@@ -1,12 +1,12 @@
-import EccUtilsDesignForm from "./form.js";
+import Form from "./form.js";
 
 export * from "./form.js";
-export default EccUtilsDesignForm;
+export default Form;
 
-window.customElements.define("ecc-utils-design-form", EccUtilsDesignForm);
+window.customElements.define("ecc-utils-design-form", Form);
 
 declare global {
   interface HTMLElementTagNameMap {
-    "ecc-utils-design-form": EccUtilsDesignForm;
+    "ecc-utils-design-form": Form;
   }
 }
