+++ packages/ecc-utils-design/src/components/index.ts
@@ -1,3 +1,3 @@
-export { default as EccUtilsDesignForm } from "./form/index.js";
-export { default as EccUtilsDesignCollection } from "./collection/index.js";
-export { default as EccUtilsDesignDetails } from "./details/details.js";
+export { default as Form } from "./form/index.js";
+export { default as Collection } from "./collection/index.js";
+export { default as Details } from "./details/details.js";
