--- packages/ecc-utils-design/src/events/ecc-utils-button-click.ts
@@ -1,7 +1,0 @@
-export type EccUtilsButtonClickEvent = CustomEvent<Record<string, any>>;
-
-declare global {
-  interface GlobalEventHandlersEventMap {
-    "ecc-utils-button-click": EccUtilsButtonClickEvent;
-  }
-}
