--- packages/ecc-utils-design/src/events/ecc-utils-expand.ts
@@ -1,7 +1,0 @@
-export type EccUtilsExpandEvent = CustomEvent<{ key?: number }>;
-
-declare global {
-  interface GlobalEventHandlersEventMap {
-    "ecc-utils-expand": EccUtilsExpandEvent;
-  }
-}
