--- packages/ecc-utils-design/src/events/ecc-utils-filter.ts
@@ -1,7 +1,0 @@
-export type EccUtilsFilterEvent = CustomEvent<{ key: string; value: string }>;
-
-declare global {
-  interface GlobalEventHandlersEventMap {
-    "ecc-utils-filter": EccUtilsFilterEvent;
-  }
-}
