--- packages/ecc-utils-design/src/events/ecc-utils-page-change.ts
@@ -1,7 +1,0 @@
-export type EccUtilsPageChangeEvent = CustomEvent<{ page: number }>;
-
-declare global {
-  interface GlobalEventHandlersEventMap {
-    "ecc-utils-page-change": EccUtilsPageChangeEvent;
-  }
-}
