--- packages/ecc-utils-design/src/events/ecc-utils-submit.ts
@@ -1,7 +1,0 @@
-export type EccUtilsSubmitEvent = CustomEvent<{ form: object }>;
-
-declare global {
-  interface GlobalEventHandlersEventMap {
-    "ecc-utils-submit": EccUtilsSubmitEvent;
-  }
-}
