--- packages/ecc-utils-design/src/events/index.ts
@@ -1,5 +1,0 @@
-export type { EccUtilsSubmitEvent } from "./ecc-utils-submit.js";
-export type { EccUtilsExpandEvent } from "./ecc-utils-expand.js";
-export type { EccUtilsFilterEvent } from "./ecc-utils-filter.js";
-export type { EccUtilsPageChangeEvent } from "./ecc-utils-page-change.js";
-export type { EccUtilsButtonClickEvent } from "./ecc-utils-button-click.js";
