+++ packages/ecc-utils-design/src/index.ts
@@ -1,3 +1,1 @@
 import "./components/index.js";
-import "./events/index.js";
-import "./utilities/index.js";
